Implement unit of measure and conversion
A product has one base unit of measure and that unit
has a list of conversion factors that records how
to convert that base unit to another unit.

For example, Product A has a base unit of `piece`.
To convert `piece` to `box`, it needs 12 `pieces`,
which is the factor.

Include test cases.

Signed-off-by: Joan Tiffany Siy <joan@teamcodeflux.com>